tickets ui update

+++ lib/models/events/posts/make_request_model.dart
@@ -5,7 +5,7 @@ class MakeRequestModel {
   int members;
   int userId;
   bool userDecision;
-  int totalAmount;
+  double totalAmount;
 
   MakeRequestModel({
     required this.userId,
